refactor: addDevServer
