+++ server/index.js
@@ -1,44 +1,17 @@
 const express = require('express');
-const { clientBuild, serverBuild, statsFilename } = require('./constants');
+const { clientBuild } = require('./constants');
 const render = require('./render').default;
+const addDevServer = require('./middlewares/addDevServer');
 
 const app = express();
 const host = '127.0.0.1';
 const port = 3000;
 
-if (process.env.NODE_ENV !== 'production') {
-  /* eslint-disable global-require, import/no-extraneous-dependencies */
-  const webpack = require('webpack');
-  const webpackDevMiddleware = require('webpack-dev-middleware');
-  const webpackHotMiddleware = require('webpack-hot-middleware');
-  const [
-    clientConfig,
-    serverConfig,
-  ] = require('../internals/webpack/dev.config');
-  /* eslint-enable global-require, import/no-extraneous-dependencies */
-
-  const compiler = webpack([clientConfig, serverConfig]);
-  const clientCompiler = compiler.compilers.find(
-    ({ name }) => name === 'client'
-  );
-  const webpackDevMiddlewareOptions = {
-    logLevel: 'error',
-    publicPath: clientConfig.output.publicPath,
-    watchOptions: {
-      ignored: /node_modules/,
-    },
-    writeToDisk: filePath =>
-      new RegExp(serverBuild).test(filePath) ||
-      new RegExp(statsFilename).test(filePath),
-  };
-
-  app.use(webpackDevMiddleware(compiler, webpackDevMiddlewareOptions));
-
-  // only use clientCompiler, otherwise hot reload would fail
-  app.use(webpackHotMiddleware(clientCompiler));
-}
-
 app.use(express.static(clientBuild));
+
+if (process.env.NODE_ENV !== 'production') {
+  addDevServer(app);
+}
 
 app.get('*', (req, res) => {
   const { url } = req;
